isolate custom trigger and registration logics
