+++ core/src/main/java/io/github/wysohn/triggerreactor/core/IEventHook.java
@@ -1,0 +1,6 @@
+package io.github.wysohn.triggerreactor.core;
+
+@FunctionalInterface
+public interface IEventHook {
+    void onEvent(Object e);
+}
